@@ -1,34 +1,36 @@
 package com.jwt.ws.api.security;
 
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationConverter;
+import org.springframework.security.web.SecurityFilterChain;
 
-@EnableGlobalMethodSecurity(securedEnabled=true, prePostEnabled=true)
+@Configuration
+@EnableMethodSecurity(securedEnabled=true, prePostEnabled=true)
 @EnableWebSecurity
-public class WebSecurity extends WebSecurityConfigurerAdapter {
+public class WebSecurity {
 	
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
+	@Bean
+	public SecurityFilterChain configure(HttpSecurity http) throws Exception {
 		
 		JwtAuthenticationConverter jwtAuthenticationConverter = new JwtAuthenticationConverter();
 		jwtAuthenticationConverter.setJwtGrantedAuthoritiesConverter(new KeycloakRoleConverter());
-	 
-		http
-			.authorizeRequests()
-					.antMatchers(HttpMethod.GET, "/users/status/check") 
-					//.hasAuthority("SCOPE_profile")
-					.hasRole("developer")
-					//.hasAnyAuthority("ROLE_developer")
-					//.hasAnyRole("devleoper","user")
-				.anyRequest().authenticated()
-				.and()
-			.oauth2ResourceServer()
-			.jwt()
-			.jwtAuthenticationConverter(jwtAuthenticationConverter);
+ 
+		
+		http 
+		.authorizeHttpRequests((authz) -> authz.requestMatchers(HttpMethod.GET, "/users/status/check")
+				// .hasAuthority("SCOPE_profile")
+				.hasRole("developer")
+				// .hasAnyAuthority("ROLE_developer")
+				// .hasAnyRole("developer","user")
+				.anyRequest().authenticated())
+		.oauth2ResourceServer(
+				oauth2 -> oauth2.jwt(jwt -> jwt.jwtAuthenticationConverter(jwtAuthenticationConverter)));
+		
+		return http.build();
 	}
-	
 }
